apps/it_department/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "it-department"
version = "0.1.0"
description = "CrewAI-based IT department orchestrator with Ollama and Telegram progress updates."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "crewai>=0.121.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "requests>=2.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0"
]

[project.scripts]
it-department = "it_department.main:cli"
it-department-telegram = "it_department.telegram_bot:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
